Fix some out-of-box build errors by snipping in small chunks of sdc_sdk.h
We have exposed some constants from our sdc_sdk.h in a few spots of our
API. These will be replaced by our schema constants later, but for now
we'll need these for building. So we don't ask our customers to plug-in
the full sdc_sdk.h, we'll include the full small chunks that are required.