build: fixups for position-independent library changes
The dcal library should be a complete package and be able to be included
without including other libraries (unless the code directly uses other
libraries, like our tests make extra use of libssh). This facilitates
using it within other wrappers and projects. So, change the way we build
it to include libssh and libflatccrt directly within libdcal itself. Adjust
also for a -fPIC build.

Had to adjust the libflatcc build to add -fPIC through the proper cmake
option. This adds a patch to do so.

After changing the way the library was built, had to adjust the way our
various test and example code was built. Notice that I was unable to
remove libssh as a dependency there because our common code that's not
part of libdcal independantly uses libssh.